+++ stash/stash_engine/lib/stash/zenodo_replicate/deposit.rb
@@ -1,25 +1,12 @@
-require 'http'
-
-# require 'stash/zenodo_replicate'
-# resource = StashEngine::Resource.find(785)
-# z = Stash::ZenodoReplicate::ZenodoConnection.new(resource: resource, file_collection:)
-# The zenodo newversion seems to be editing the same deposition id
-# 503933
-
 module Stash
   module ZenodoReplicate
-
-    class ZenodoError < StandardError; end
-
-    class ZenodoConnection
+    class Deposit
 
       attr_reader :resource, :file_collection, :deposit_id, :links, :files
 
       def initialize(resource:, file_collection:)
         @resource = resource
         @file_collection = file_collection
-
-        @http = HTTP.timeout(connect: 30, read: 60).timeout(7200).follow(max_hops: 10)
       end
 
       # checks that can access API with token and return boolean
@@ -80,67 +67,9 @@ module Stash
         resp
       end
 
-      def delete_files
-        resp = standard_request(:get, "#{base_url}/api/deposit/depositions/#{deposit_id}")
-
-        resp[:files].map do |f|
-          standard_request(:delete, f[:links][:download])
-        end
-
-        @files = [] # now it's empty
-
-        standard_request(:get, "#{base_url}/api/deposit/depositions/#{deposit_id}")
-      end
-
-      def send_files
-        path = @file_collection.path.to_s
-        path << '/' unless path.end_with?('/')
-
-        all_files = Dir["#{path}/**/*"]
-
-        all_files.each do |f|
-          short_fn = f[path.length..-1]
-          resp = standard_request(:put, "#{links[:bucket]}/#{ERB::Util.url_encode(short_fn)}", body: File.open(f, 'rb'))
-
-          # TODO: check the response digest against the known digest
-        end
-      end
-
-      def get_files_info
-        # right now this is mostly just used for internal testing
-        standard_request(:get, links[:bucket])
-      end
-
       def publish
         standard_request(:post, links[:publish])
       end
-
-      private
-
-      def standard_request(method, url, **args)
-        my_params = { access_token: APP_CONFIG[:zenodo][:access_token] }.merge(args.fetch(:params, {}))
-        my_headers = { 'Content-Type': 'application/json' }.merge(args.fetch(:headers, {}))
-        my_args = args.merge(params: my_params, headers: my_headers)
-
-        r = @http.send(method, url, my_args)
-
-        resp = r.parse
-        resp = resp.with_indifferent_access if resp.class == Hash
-
-        unless r.status.success?
-          raise ZenodoError, "Zenodo response: #{r.status.code}\n#{resp} for \nhttp.#{method} #{url}\n#{resp}"
-        end
-        resp
-      end
-
-      def param_merge(p = {})
-        { access_token: APP_CONFIG[:zenodo][:access_token] }.merge(p)
-      end
-
-      def base_url
-        APP_CONFIG[:zenodo][:base_url]
-      end
-
     end
   end
 end
